Allow disconnected-in-time versions
This removes the restriction that v1's version_end_date
must equal v2's version_start_date.

Doing so will allow (with code in future commits) a deleted
version to be re-activated at a later point in time.  For example:

    t1: v1 created
    t2: v1 deleted
    t3: v2 restored

In this case, v1 would look like this:
    identity=abc, version_birth_date=t1, version_start_date=t1, version_end_date=t2
And v2 would look like this:
    identity=abc, version_birth_date=t1, version_start_date=t3, version_end_date=None

It is still possible to navigate versions with previous_version,
next_version.